# api-gateway/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: BoShop API
  description: Полноценный e-commerce на микросервисах
  version: 1.0.0
servers:
  - url: /
    description: API Gateway
...
paths:
...
  /api/v1/users/me:
    get:
      tags:
        - Users
      summary: Получить текущего пользователя (авторизованный)
      security:
        - bearerAuth: [ROLE_USER, ROLE_ADMIN]
      responses:
        '200':
          description: Приветствие
          content:
            application/json:
              schema:
                type: string
  /api/v1/users:
    get:
      tags:
        - Users
      summary: Получить всех пользователей (только ADMIN)
      security:
        - bearerAuth: [ROLE_ADMIN]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedModelUserDto'

    post:
      tags:
        - Users
      summary: Создать пользователя (ADMIN)
      security:
        - bearerAuth: [ROLE_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        '200':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
  /api/v1/users/{id}:
    get:
      tags:
        - Users
      summary: Получить пользователя по ID (ADMIN)
      security:
        - bearerAuth: [ROLE_ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'

    patch:
      tags:
        - Users
      summary: Частичное обновление пользователя (ADMIN)
      security:
        - bearerAuth: [ROLE_ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        '200':
          description: Пользователь обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'

    delete:
      tags:
        - Users
      summary: Удалить пользователя (ADMIN)
      security:
        - bearerAuth: [ROLE_ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Пользователь удалён
...
  /api/v1/products/public/search:
    get:
      tags:
        - Products
      summary: Поиск и фильтрация товаров (публичный)
      security: []
      parameters:
        - name: name
          in: query
          schema:
            type: string
          description: Название товара (частичное совпадение)
        - name: category
          in: query
          schema:
            type: string
          description: Категория товара
        - name: published
          in: query
          schema:
            type: boolean
            default: true
          description: Фильтр по статусу публикации
        - name: priceFrom
          in: query
          schema:
            type: number
            format: double
          description: Минимальная цена
        - name: priceTo
          in: query
          schema:
            type: number
            format: double
          description: Максимальная цена
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Страница с товарами
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageProductResponseDto'
  /api/v1/products/public/{id}:
    get:
      tags:
        - Products
      summary: Получить товар по ID (публичный)
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Данные товара
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDto'
        '404':
          description: Товар не найден
    # === INVENTORY (USER) ===


  # === INVENTORY (ADMIN) ===
...
  /api/v1/orders/public/{id}:
    get:
      tags:
        - Orders
      summary: Получить заказ по ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Данные заказа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponseDto'
        '404':
          description: Заказ не найден
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductRequestDto:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          maxLength: 255
          example: "Wireless Headphones"
        description:
          type: string
          maxLength: 1000
          example: "High-quality wireless headphones with noise cancellation"
        price:
          type: number
          format: double
          minimum: 0.01
          example: 99.99
        category:
          type: string
          maxLength: 50
          example: "Electronics"
        published:
          type: boolean
          default: false
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 10
          example: [ "audio", "wireless", "gadget" ]
        images:
          type: array
          items:
            type: string
            format: uri
          maxItems: 5
          example: [ "https://example.com/images/headphones1.jpg" ]
    ProductResponseDto:
      allOf:
        - $ref: '#/components/schemas/ProductRequestDto'
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    OrderResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 456
        userId:
          type: integer
          format: int64
          example: 123
        status:
          type: string
          enum: [ NEW, RESERVED, PAID, CANCELLED ]
          example: NEW
        totalPrice:
          type: number
          format: double
          example: 199.98
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PageProductResponseDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponseDto'
        page:
          type: object
          properties:
            size:
              type: integer
            totalElements:
              type: integer
            totalPages:
              type: integer
            number:
              type: integer
...
    UserDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    PagedModelUserDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
        page:
          type: object
          properties:
            size:
              type: integer
            totalElements:
              type: integer
            totalPages:
              type: integer
            number:
              type: integer